requirements.txt
python-dotenv
langgraph "langchain[anthropic]"
langgraph langsmith
pydantic